requirements.txt
selenium==3.141.0
urllib3==1.22
behave==1.2.6
requests==2.26.0
python-dateutil==2.8.0
browserstack-local==1.2.3